+++ simon-plaintext-test.cpp
@@ -21,9 +21,9 @@ int main(int argc, char **argv)
     printKey(k);
     
     vector<clock_t> times;
-    for (int i = 0; i < 1000; i++) {
+    for (size_t i = 0; i < 1000; i++) {
         clock_t t = clock();
-        for (int j = 0; j < T; j++) {
+        for (size_t j = 0; j < T; j++) {
             bs[0] = pt_encRound(k[j], bs[0]);
         }
         t = clock() - t;
